apps/api/pyproject.toml
[project]
name = "ai-digital-human-api"
version = "0.1.0"
description = "FastAPI backend for AI Digital Human system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "qdrant-client>=1.7.0",
    "insightface>=0.7.3",
    "faster-whisper>=0.10.0",
    "openai>=1.12.0",
    "aisuite>=0.1.0",
    "tiktoken>=0.7.0",
    "pillow>=10.2.0",
    "opencv-python-headless>=4.9.0",
    "aiosqlite>=0.19.0",
    "fastmcp>=0.1.0",
    "zhipuai>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
